+++ module/genesis_test.go
@@ -144,7 +144,7 @@ func (suite *GenesisTestSuite) TestGenesisState() {
 				randomNumber, _ := bep3.GenerateSecureRandomNumber()
 				randomNumberHash := bep3.CalculateRandomHash(randomNumber[:], timestamp)
 				swap := bep3.NewAtomicSwap(cs(c("bnb", halfLimit)), randomNumberHash,
-					uint64(360), timestamp, suite.addrs[0], addrs[1], TestSenderOtherChain,
+					360, timestamp, suite.addrs[0], addrs[1], TestSenderOtherChain,
 					TestRecipientOtherChain, 0, bep3.Open, true, bep3.Incoming)
 				gs.AtomicSwaps = bep3.AtomicSwaps{swap}
 
@@ -214,7 +214,7 @@ func (suite *GenesisTestSuite) TestGenesisState() {
 				randomNumber, _ := bep3.GenerateSecureRandomNumber()
 				randomNumberHash := bep3.CalculateRandomHash(randomNumber[:], timestamp)
 				swap := bep3.NewAtomicSwap(cs(c("fake", 500000)), randomNumberHash,
-					uint64(360), timestamp, suite.addrs[0], addrs[1], TestSenderOtherChain,
+					360, timestamp, suite.addrs[0], addrs[1], TestSenderOtherChain,
 					TestRecipientOtherChain, 0, bep3.Open, true, bep3.Incoming)
 
 				gs.AtomicSwaps = bep3.AtomicSwaps{swap}
@@ -231,7 +231,7 @@ func (suite *GenesisTestSuite) TestGenesisState() {
 				randomNumber, _ := bep3.GenerateSecureRandomNumber()
 				randomNumberHash := bep3.CalculateRandomHash(randomNumber[:], timestamp)
 				swap := bep3.NewAtomicSwap(cs(c("bnb", 5000)), randomNumberHash,
-					uint64(360), timestamp, suite.addrs[0], addrs[1], TestSenderOtherChain,
+					360, timestamp, suite.addrs[0], addrs[1], TestSenderOtherChain,
 					TestRecipientOtherChain, 0, bep3.NULL, true, bep3.Incoming)
 
 				gs.AtomicSwaps = bep3.AtomicSwaps{swap}
